Move test to CParserTests
The behavior is only applicable on the `c` engine.
