+++ pandas/tests/io/parser/comment.py
@@ -5,8 +5,6 @@ Tests that comments are properly handled during parsing
 for all of the parsers defined in parsers.py
 """
 
-import sys
-
 import numpy as np
 import pandas.util.testing as tm
 
@@ -118,26 +116,3 @@ A,B,C
         expected = DataFrame({0: ['a', '1'], 1: ['b', '2'], 2: ['c', '3']})
         result = self.read_csv(StringIO(data), comment='#', header=None)
         tm.assert_frame_equal(result, expected)
-
-    @tm.capture_stderr
-    def test_comment_whitespace_delimited(self):
-        test_input = """\
-1 2
-2 2 3
-3 2 3 # 3 fields
-4 2 3# 3 fields
-5 2 # 2 fields
-6 2# 2 fields
-7 # 1 field, NaN
-8# 1 field, NaN
-9 2 3 # skipped line
-# comment"""
-        df = self.read_csv(StringIO(test_input), comment='#', header=None,
-                           delimiter='\\s+', skiprows=0,
-                           error_bad_lines=False)
-        error = sys.stderr.getvalue()
-        # skipped lines 2, 3, 4, 9
-        for line_num in (2, 3, 4, 9):
-            assert 'Skipping line {}'.format(line_num) in error, error
-        expected = DataFrame([[1, 2], [5, 2], [6, 2], [7, np.nan], [8, np.nan]])
-        tm.assert_frame_equal(df, expected)
